requirements.txt
hickle==3.4.3
six==1.12.0
tensorflow_gpu==1.12.0
scipy==1.0.1
np_utils==0.5.10.0
pandas==0.24.2
stft==0.5.2
numpy==1.11.0
mne==0.11.0
scikit_learn==0.21.3
